bugfixes, simplified output for penalty algorithm, fileio_append
git-svn-id: svn://localhost/trunk@1060 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/surfit/variables.cpp
@@ -62,7 +62,7 @@ int reproject_faults = 1;
 int reproject_undef_areas = 0;
 int process_isolated_areas = 1;
 
-int penalty_max_iter = 150;
+size_t penalty_max_iter = 150;
 REAL penalty_weight = 1;
 REAL penalty_weight_mult = 2;
