+++ surfit/src/surfit/variables_tcl.h
@@ -84,7 +84,7 @@ namespace surfit {
 	*/
 	extern SURFIT_EXPORT int process_isolated_areas;
 
-	extern int penalty_max_iter;
+	extern size_t penalty_max_iter;
 	extern REAL penalty_weight;
 	extern REAL penalty_weight_mult;
